+++ .idea/.idea.TeamZ/.idea/.gitignore
@@ -1,0 +1,13 @@
+# デフォルトの無視対象ファイル
+/shelf/
+/workspace.xml
+# Rider の無視対象ファイル
+/modules.xml
+/.idea.TeamZ.iml
+/contentModel.xml
+/projectSettingsUpdater.xml
+# エディターベースの HTTP クライアントリクエスト
+/httpRequests/
+# Datasource local storage ignored files
+/dataSources/
+/dataSources.local.xml
