+++ .idea/.idea.TeamZ/.idea/encodings.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/Assets/Scenes/Noda/ShopUpgradeItem.cs" charset="windows-31j" />
+    <file url="file://$PROJECT_DIR$/Assets/Scenes/ResourceManager.cs" charset="windows-31j" />
+  </component>
+</project>
