+++ Assets/Prefabs.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 3d1dc6cc25dc9cc46950037602a6f554
+guid: aad5338cc71722648b608812f93bc267
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
