+++ Assets/Prefabs/Canvas.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 10b419961950a52458d2508eee408ddf
-folderAsset: yes
-DefaultImporter:
+guid: eb300aabb6fcb684780222dcfe0d75ed
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
