--- Assets/SaitoFile.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0bbdf9bc2eb545f4f826f81a793cdd54
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
