+++ Assets/Scripts/ResourceManager.cs.meta
