+++ ProjectSettings/TagManager.asset
@@ -41,3 +41,9 @@ TagManager:
   - name: Default
     uniqueID: 0
     locked: 0
+  - name: Factory
+    uniqueID: 2324576011
+    locked: 0
+  - name: FactoryList
+    uniqueID: 441314641
+    locked: 0
