+++ UserSettings/EditorUserSettings.asset
@@ -6,11 +6,13 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-<<<<<<< HEAD
       value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
-=======
-      value: 5055050206030a585f5b082149205c44404e4d7a7c7b71667d7b1c65e4b5616c
->>>>>>> 105e8edf871e7ce81836ffef03fd34897250c7d3
+      flags: 0
+    RecentlyUsedSceneGuid-1:
+      value: 5757515f5d545b035d5f5d7613205d4445154e29287d73682f704562b2b6376c
+      flags: 0
+    RecentlyUsedSceneGuid-2:
+      value: 050406505c06595a080a557a40200d44474f1e2b297020352b781837e4b1613c
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
